Trick g-ir-scanner some more
Add casts to get some defined constants to have
the right type in the gir. The games we play...

+++ pango/pango-attributes.h
@@ -267,7 +267,7 @@ typedef enum {
  *
  * Since: 1.24
  */
-#define PANGO_ATTR_INDEX_FROM_TEXT_BEGINNING 0
+#define PANGO_ATTR_INDEX_FROM_TEXT_BEGINNING ((guint)0)
 
 /**
  * PANGO_ATTR_INDEX_TO_TEXT_END: (value 4294967295)
@@ -277,7 +277,7 @@ typedef enum {
  *
  * Since: 1.24
  */
-#define PANGO_ATTR_INDEX_TO_TEXT_END (G_MAXUINT + 0)
+#define PANGO_ATTR_INDEX_TO_TEXT_END ((guint)(G_MAXUINT + 0))
 
 /**
  * PangoAttribute:
